refactor(meeting): extract pump task body into peer module (#431)
Lifts run_pump + PumpContext + TickBucket +
diarize_and_dispatch_merged + dispatch_utterances out of
manager.rs into a new src/meeting/pump.rs peer. PUMP_TICK and
CANONICAL_FORMAT travel with the pump (they're only ever used
there). manager.rs shrinks 2,749 → 2,207 LOC; the pump is the
heaviest single concentration of logic in the meeting module so
extracting it is the most leverage-per-byte split.

No behaviour change. Visibility is `pub(super)` for the moved
items so manager.rs's existing tests still drive them directly
without exposing the pump's internals on the crate surface.

Refs #431.
